moved Activate() to steplib.go
